# src/api/garden-backend/src/doc/swagger.json
{
    "openapi": "3.0.1",
    "info": {
      "title": "ClassOverflowWebAPI",
      "version": "v1"
    },
    "servers": [
        {
          "url": "https://puc-garden-backend.herokuapp.com/api/v1/",
          "description": "API em Produção"
        },
        {
          "url": "https://the-class-overflow.herokuapp.com/",
          "description": "Aplicação WEB"
        }
    ],
    "paths": {
      "/posts": {
        "get": {
          "tags": [
            "Post"
          ],
          "responses": {
            "200": {
              "description": "Success",
              "content": {
                "text/plain": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Post"
                    }
                  }
                },
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Post"
                    }
                  }
                },
                "text/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Post"
                    }
                  }
                }
              }
            },
            "400": {
              "description": "Bad Request"
            },
            "500": {
              "description": "Server Error"
            }
          }
        },
        "post": {
          "tags": [
            "Post"
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Post"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Post"
                }
              },
              "application/*+json": {
                "schema": {
                  "$ref": "#/components/schemas/Post"
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "Success",
              "content": {
                "text/plain": {
                  "schema": {
                    "$ref": "#/components/schemas/Post"
                  }
                },
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Post"
                  }
                },
                "text/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Post"
                  }
                }
              }
            },
            "400": {
              "description": "Bad Request"
            }
          }
        }
      },
      "/posts/{id}": {
        "get": {
          "tags": [
            "Post"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string",
                "format": "uuid"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Success",
              "content": {
                "text/plain": {
                  "schema": {
                    "$ref": "#/components/schemas/Post"
                  }
                },
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Post"
                  }
                },
                "text/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Post"
                  }
                }
              }
            },
            "404": {
              "description": "Not found"
            },
            "500": {
              "description": "Server error"
            }
          }
        },
        "put": {
          "tags": [
            "Post"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string",
                "format": "uuid"
              }
            }
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Post"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Post"
                }
              },
              "application/*+json": {
                "schema": {
                  "$ref": "#/components/schemas/Post"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Success"
            },
            "400": {
              "description": "Bad Request"
            },
            "403": {
              "description": "only the poster owner can edit"
            },
            "404": {
              "description": "post not found"
            }
          }
        },
        "delete": {
          "tags": [
            "Post"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string",
                "format": "uuid"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Success"
            },
            "400": {
              "description": "Bad Request"
            }
          }
        }
      },
      "/posts/{id}/comments": {
        "post": {
          "tags": [
            "Comments"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Post"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Post"
                }
              },
              "application/*+json": {
                "schema": {
                  "$ref": "#/components/schemas/Post"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Success",
              "content": {
                "text/plain": {
                  "schema": {
                    "$ref": "#/components/schemas/Post"
                  }
                },
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Post"
                  }
                },
                "text/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Post"
                  }
                }
              }
            },
            "400": {
              "description": "Bad Request"
            },
            "404": {
              "description": "Not found"
            }
          }
        }
      },
      "/posts/{id}/comments/like": {
        "put": {
          "tags": [
            "Comments"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Post"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Post"
                }
              },
              "application/*+json": {
                "schema": {
                  "$ref": "#/components/schemas/Post"
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "Success",
              "content": {
                "text/plain": {
                  "schema": {
                    "$ref": "#/components/schemas/Post"
                  }
                },
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Post"
                  }
                },
                "text/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Post"
                  }
                }
              }
            },
            "204": {
              "description": "Post not found"
            },
            "400": {
              "description": "Bad Request"
            },
            "404": {
              "description": "Not found"
            }
          }
        }
      },
      "/auth":{
        "post": {
            "tags": [
              "Authorization"
            ],
            "requestBody": {
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/User"
                  }
                },
                "text/json": {
                  "schema": {
                    "$ref": "#/components/schemas/User"
                  }
                },
                "application/*+json": {
                  "schema": {
                    "User": {
                      
                    }
                  }
                }
              }
            },
            "responses": {
              "200": {
                "description": "Success",
                "content": {
                  "text/plain": {
                    "schema": {
                      "$ref": "#/components/schemas/Token"
                    }
                  },
                  "application/json": {
                    "schema": {
                      "$ref": "#/components/schemas/Token"
                    }
                  },
                  "text/json": {
                    "schema": {
                      "$ref": "#/components/schemas/Token"
                    }
                  }
                }
              },
              "400": {
                "description": "Bad Request"
              }
            }
          }
      }
    },
    "components": {
      "schemas": {
        "Token":{
            "type": "object",
            "properties": {
              "bearer":{
                "type": "string"
              }
            },
            "example":{
              "bearer": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiI1ZWJhYzUzNDk1NGI1NDEzOTgwNmMxMTIiLCJpYXQiOjE1ODkyOTg0ODQsImV4cCI6MTU4OTMwMDI4NH0.m1U63blB0MLej_WfB7yC2FTMnCziif9X8yzwDEfJXAg"
            }
        },
        "Likes":{
            "type": "object",
            "properties": {
                "owner":{
                    "type": "string"
                },
                "ownerName":{
                    "type": "string"
                },
                "isLiked": {
                    "type": "boolean"
                }
            }
        },
        "Comments": {
          "type": "object",
          "properties": {
            "id": {
              "type": "string",
              "format": "uuid"
            },
            "owner":{
                "type": "string"
            },
            "ownerName":{
                "type": "string"
            },
            "ownerEmail":{
                "type": "string"
            },
            "content": {
              "type": "string",
              "nullable": true
            },
            "createdAt": {
              "type": "string",
              "format": "date-time"
            },
            "likes":{
               "type": "array",
               "items": {
                "$ref": "#/components/schemas/Likes"
               },
               "nullable": true
            }
          },
          "additionalProperties": false
        },
        "Post": {
          "type": "object",
          "properties": {
            "_id": {
              "type": "string",
              "format": "uuid"
            },
            "title": {
              "type": "string",
              "nullable": true
            },
            "content": {
              "type": "string",
              "nullable": true
            },
            "isActive": {
              "type": "boolean"
            },
            "createdAt": {
              "type": "string",
              "format": "date-time"
            },
            "updatedAt": {
              "type": "string",
              "format": "date-time"
            },
            "topic":{
              "type": "array"
            },
            "owner":{
                "type": "string"
            },
            "ownerName":{
                "type": "string"
            },
            "ownerEmail":{
                "type": "string"
            },
            "comments": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/Comments"
              },
              "nullable": true
            }
          },
          "additionalProperties": false
        },
        "User": {
          "type": "object",
          "properties": {
            "name": {
              "type": "string",
              "nullable": true
            },
            "password": {
              "type": "string",
              "nullable": true
            }
          },
          "additionalProperties": false
        }
      },
      "securitySchemes":{
          "bearerAuth":{
            "type": "http",
            "scheme": "bearer",
            "bearerFormat": "JWT"
          }
      }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ]
  }